project/requirements.txt
streamlit==1.31.1
mysql-connector-python==8.3.0
python-dotenv==1.0.1
bcrypt==4.1.2
timeago==1.0.16
Pillow==10.2.0